Remove seuil non-versement PPA in definition of PPA éligibilité when defining take-up
For very small amounts of revenus d'activité, this caused eligibility to PPA to be set to False because amounts of PPA were inferior to the seuil de non-versement.
This may not have caused problems when applying to data, but resulted in no PPA *and* no RSA Socle for very low earnings in cas-types.